requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-api-python-client
httpx
openai
pydantic
redis
requests
boto3
python-multipart
opencv-python
ultralytics
openai-whisper
pgvector
sentence-transformers
celery
dependency_injector
aiohttp
aiofiles
yt_dlp
asyncpg